backend/pag/serializers.py
from rest_framework import serializers
from .models import UserProfile, Transaction

class UserProfileSerializer(serializers.ModelSerializer):
    class Meta:
        model = UserProfile
        fields = '__all__'
